+++ src/Randomizer/MapMusicObject.js
@@ -11,7 +11,7 @@ class MapMusicObject extends GridTableObject {
 
     static get validMusics() {
         if(this._validMusics !== undefined) return this._validMusics;
-        this._validMusics = new Set([4, 13, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 77, 81, 83, 84, 86, 87, 90, 91, 93, 95, 106, 107, 108, 112, 114, 116, 117, 119, 120, 121, 122, 125, 128, 129, 130, 131, 132, 135, 136, 140, 142, 144, 146, 147, 150, 151, 152, 153, 154, 156, 159, 169, 170, 171, 173, 178, 187, 188]);
+        this._validMusics = new Set([2, 3, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 76, 77, 78, 80, 81, 82, 83, 84, 86, 87, 90, 91, 92, 106, 107, 108, 112, 114, 116, 117, 118, 119, 120, 121, 122, 125, 128, 129, 130, 131, 132, 135, 136, 140, 142, 144, 146, 147, 149, 150, 151, 152, 153, 154, 159, 169, 170, 171, 173, 178, 187, 188]);
         if(this.context.specs.flags.w >= 3) {
             EnemyObject.every.forEach(e => this._validMusics.add(e.oldData.music));
         }
@@ -24,6 +24,12 @@ class MapMusicObject extends GridTableObject {
         this._battleMusics = new Set(EnemyObject.every.map(e => e.oldData.music));
         return this.battleMusics;
     }
+
+    static get clearFloorFlags() {
+        return [...Array(9).keys()].map(caveLevelZeroIndex =>
+            [0x05, this.flagBase + caveLevelZeroIndex, 0x03]
+        );
+    }
     
     static mutateAll() {
         this.classReseed("mut");
@@ -33,8 +39,6 @@ class MapMusicObject extends GridTableObject {
             EnemyObject.every.forEach(e => e.data.music = this.context.random.choice(musics));
         }
 
-        // Hack map music to not use lookup table
-        this.context.rom.set([0x80, 0x6B], 0x6928);
 
         if(this.context.specs.flags.a) {
             // Map music shuffle - Ancient Cave mode
@@ -43,14 +47,25 @@ class MapMusicObject extends GridTableObject {
                 if(meo.caveLevel === null || meo.caveLevel === undefined) return;
                 const mmo = meo.music;
                 if(mmo.mutated) return;
-                mmo.data.music_index = chosenMusics[meo.caveLevel - 1];
+                mmo.data.music_index = 2;
                 mmo.mutated = true;
             });
-            // Ignore sector changes for music
-            this.context.rom.set([0x80], 0x5277);
+
+            // Replace second entry in the overworld_event_music_table with our new music flag data
+            const newMusicData = [];
+            chosenMusics.forEach((chosenMusic, index) => {
+                newMusicData.push(...[(this.flagBase + index), 0x83, chosenMusic, 0x00]);
+            });
+
+            newMusicData.push(...[0x00, 0x00, 156, 0x00]); // Meteor fall
+            this.context.rom.set(newMusicData, 0xF5A3D);
         }
         else {
             // Map music shuffle - Non-Ancient Cave mode
+
+            // Hack map music to not use lookup table
+            this.context.rom.set([0x80, 0x6B], 0x6928);
+
             const originalMusics = Array.from(this.validMusics);
             const chosenMusics = this.context.random.shuffle(Array.from(this.validMusics));
             const assign = {};
@@ -63,6 +78,8 @@ class MapMusicObject extends GridTableObject {
         }
     }
 }
+
+MapMusicObject.flagBase = 0xE0; // 0x3e0
 
 MapMusicObject.rows = 80;
 MapMusicObject.columns = 32;