+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v41.x / v42 (In progress)
+- Music: Music selection has been updated, and Ancient Cave music randomization will be much more consistent.
+
 ## v41 (2019-06-29)
 - Character stats: Added shuffle of special abilities (Pray, Spy, Mirror). Mirror will have an increased chance of success and Spy will give more detailed information if this mode is activated.
 ### Bugfixes